requirements/development.txt
-r ./base.txt

# Testing
factory-boy==2.12.0
Faker==4.1.1
pytest-django==3.9.0
pytest-cov==2.10.0
pytest-watch==4.2.0
pytest-picked==0.4.4
django-webtest==1.9.7
time-machine==1.2.0

# Code quality
black==19.10b0
pylint-django==2.1.0
pylint-celery==0.3
coverage==5.2.1

# Debugging
django-extensions==2.2.9
django-debug-toolbar==2.2
django-debug-permissions==1.0.0
requests==2.24.0